requirements.txt
Django==3.0.5
Djangorestframework==3.11.0
